+++ src/marknotes/session.php
@@ -46,6 +46,14 @@ class Session
             self::set('marknotes', 1);
         }
 
+        $arr=$aeSettings->getAdminInfos();
+        $login= isset($arr['login']) ? trim($arr['login']) : '';
+        $password= isset($arr['password']) ? trim($arr['password']) : '';
+
+        // If both login and password are empty (will probably be the case on a localhost server),
+        // consider the user already authenticated
+        if (($login==='') && ($password==='')) self::set('authenticated', 1);
+
         return;
     }