+++ src/marknotes/tasks/display.php
@@ -28,56 +28,6 @@ class Display
         return self::$_Instance;
     } // function getInstance()
 
-    /**
-     * Display the HTML rendering of the note in a nice HTML layout. Called when the URL is something like
-     * http://localhost/notes/docs/Development/atom/Plugins.html i.e. accessing the .html file
-     *
-     * @param  string  $html [description]   html rendering of the .md file
-     * @return {[type]       Nothing
-     */
-    private function showHTML(string $html, array $params = null)
-    {
-        $aeFiles = \MarkNotes\Files::getInstance();
-        $aeHTML = \MarkNotes\FileType\HTML::getInstance();
-        $aeSettings = \MarkNotes\Settings::getInstance();
-
-        // Add h2 and h3 id and don't add the "go to top" icon
-        $html = $aeHTML->addHeadingsID($html, false);
-
-        // Add css to bullets
-        $html = $aeHTML->setBulletsStyle($html);
-
-        // Check if a template has been specified in the parameters
-        // and if so, check that this file exists
-
-        // Default is html
-        $template = $aeSettings->getTemplateFile('html');
-
-        if (isset($params['template'])) {
-            $template = $aeSettings->getTemplateFile($params['template']);
-            if (!$aeFiles->fileExists($template)) {
-                $template = $aeSettings->getTemplateFile('html');
-            }
-        }
-
-        // Don't keep the § (tags prefix) for slideshow
-        $html = str_replace($aeSettings->getTagPrefix(), '', $html);
-        $html = str_replace(htmlentities($aeSettings->getTagPrefix()), '', $html);  // &sect; = §
-
-        if ($aeFiles->fileExists($template)) {
-            $html = $aeHTML->replaceVariables(file_get_contents($template), $html, $params);
-        }
-
-        $javascript =
-        "\nvar marknotes = {};\n".
-        "marknotes.message={};\n".
-        "marknotes.message.on_this_page='".$aeSettings->getText('on_this_page', 'On this page', true)."';\n";
-
-        $html = str_replace('<!--%ADDITIONNAL_JS%-->', '<script type="text/javascript">'.$javascript.'</script>', $html);
-
-        return $html;
-    }
-
     public function run(array $params)
     {
         $aeFiles = \MarkNotes\Files::getInstance();
@@ -128,7 +78,10 @@ class Display
 
         // Check if the .html version of the markdown file already exists; if not, create it
         if (!$aeFunctions->isAjaxRequest()) {
-            return self::showHTML($html, $params);
+
+            $aeConvert = \MarkNotes\Tasks\Converter\HTML::getInstance();
+            return $aeConvert->run($html, $params);
+
         } else { // if (!\MarkNotes\Functions::isAjaxRequest())
             // -----------------------------------------------------------------------
             // Check if the file contains words present in the tags.json file : if the file being displayed
@@ -186,7 +139,7 @@ class Display
             $aeToolbar = \MarkNotes\View\Toolbar::getInstance();
 
             $html = $aeToolbar->getToolbar($params).'<div id="icon_separator" class="only_screen"/><div id="note_content">'.$html.'</div>';
-            
+
             $html = '<div class="hidden filename">'.utf8_encode($fullname).'</div>'.$html.'<hr/>';
         } // if (!\MarkNotes\Functions::isAjaxRequest())
         return $html;